+++ lib/src/fields/smart_dropdown_field.dart
@@ -23,7 +23,7 @@ class SmartDropdownField<T> extends StatefulWidget {
     this.requiredMessage = 'This field is required.',
     this.validators = const [],
     this.asyncValidators = const [],
-    this.autovalidateMode = AutovalidateMode.onUnfocus,
+    this.autovalidateMode,
     this.errorAnimation,
     this.enabled = true,
     this.decoration = const InputDecoration(),
@@ -45,7 +45,7 @@ class SmartDropdownField<T> extends StatefulWidget {
   final String requiredMessage;
   final List<SmartValidator<T>> validators;
   final List<SmartAsyncValidator<T>> asyncValidators;
-  final AutovalidateMode autovalidateMode;
+  final AutovalidateMode? autovalidateMode;
   final SmartErrorAnimation? errorAnimation;
   final bool enabled;
   final InputDecoration decoration;