+++ lib/src/fields/smart_email_field.dart
@@ -19,7 +19,7 @@ class SmartEmailField extends StatefulWidget {
     this.validators = const [],
     this.asyncValidators = const [],
     this.asyncValidationDebounce,
-    this.autovalidateMode = AutovalidateMode.onUnfocus,
+    this.autovalidateMode,
     this.errorAnimation,
     this.enabled = true,
     this.decoration = const InputDecoration(),
@@ -39,7 +39,7 @@ class SmartEmailField extends StatefulWidget {
   final List<SmartValidator<String>> validators;
   final List<SmartAsyncValidator<String>> asyncValidators;
   final Duration? asyncValidationDebounce;
-  final AutovalidateMode autovalidateMode;
+  final AutovalidateMode? autovalidateMode;
   final SmartErrorAnimation? errorAnimation;
   final bool enabled;
   final InputDecoration decoration;